+++ app/controllers/api/v1/concerns/guild_cacher.rb
@@ -9,7 +9,7 @@ module Api
         def get_guilds_from_cache
           guilds = Guild.order(id: :asc).includes(:fraction, :world)
           @guilds_json = Rails.cache.fetch(Guild.cache_key(guilds)) do
-            ActiveModelSerializers::SerializableResource.new(guilds, each_serializer: GuildSerializer).as_json[:guilds]
+            ActiveModelSerializers::SerializableResource.new(guilds.includes(:fraction, :world), each_serializer: GuildSerializer).as_json[:guilds]
           end
         end
       end