+++ app/interactors/create_delivery_with_params.rb
@@ -1,0 +1,6 @@
+# Create delivery and delivery param
+class CreateDeliveryWithParams
+  include Interactor::Organizer
+
+  organize CreateDelivery, CreateDeliveryParam
+end