+++ app/serializers/delivery_serializer.rb
@@ -1,0 +1,8 @@
+class DeliverySerializer < ActiveModel::Serializer
+  attributes :id, :deliveriable
+
+  def deliveriable
+    return GuildIndexSerializer.new(object.deliveriable) if object.deliveriable_type == 'Guild'
+    nil
+  end
+end