+++ app/serializers/guild_index_serializer.rb
@@ -1,3 +1,3 @@
 class GuildIndexSerializer < ActiveModel::Serializer
-  attributes :id, :full_name
+  attributes :id, :full_name, :name, :slug
 end