+++ app/serializers/notification_serializer.rb
@@ -1,0 +1,3 @@
+class NotificationSerializer < ActiveModel::Serializer
+  attributes :id, :name, :event
+end