+++ frontend/components/guild_management/guild_management.js
@@ -1,0 +1,1 @@
+import "./guild_management.scss";