+++ frontend/packs/delivery_form.jsx
@@ -1,0 +1,4 @@
+import DeliveryForm from "components_react/delivery_form/delivery_form";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({DeliveryForm});