+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Notification model
 - Delivery model
 - DeliveryParam model
+- creating delivery with param
 
 ## [0.6.2] - 2019-11-04
 ### Added